+++ changelog.md
@@ -7,6 +7,7 @@ This version is not released yet and is under active development.
 ```
 
 - Print all modified loggers and their levels in `DEBUG` mode.
+- Directly download in Pygments source code from GitHub  to check for candidates for ANSI-coloring in unittests.
 
 ## {gh}`4.0.0 (2023-05-08) <compare/v3.10.0...v4.0.0>`